refactor(Wikicache): optimize logic (follow #1309)
