+++ src/Wikicache/Wikicache.ts
@@ -1,19 +1,12 @@
 import * as OPTIONS from './options.json';
-import {getCache} from './modules/getCache';
-import {getCacheKey} from './modules/getCacheKey';
-import {setCache} from './modules/setCache';
+import {getCache, setCache} from './modules/cache';
 
-(function wikicache(): void {
+(async function wikicache(): Promise<void> {
 	if (mw.config.get(OPTIONS.configKey)) {
 		return;
 	}
 	mw.config.set(OPTIONS.configKey, true);
 
-	const cacheKey = getCacheKey();
-
-	void getCache(cacheKey);
-
-	void (async () => {
-		await setCache(cacheKey);
-	})();
+	await getCache();
+	await setCache();
 })();
