+++ src/Wikicache/modules/wpTextbox1Content.ts
@@ -1,12 +1,8 @@
-import {getBody} from 'ext.gadget.Util';
-
-const getWpTextbox1Content = async () => {
-	const $body = await getBody();
+const getWpTextbox1Content = ({$body}: {$body: JQuery<HTMLBodyElement>}) => {
 	return $body.find('textarea[name=wpTextbox1]').textSelection('getContents');
 };
 
-const setWpTextbox1Content = async (content: string) => {
-	const $body = await getBody();
+const setWpTextbox1Content = ({$body, content}: {$body: JQuery<HTMLBodyElement>; content: string}) => {
 	$body.find('textarea[name=wpTextbox1]').textSelection('setContents', content);
 };
